% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{fromCharToFactorNum}
\alias{fromCharToFactorNum}
\title{Convert character variable to numeric factor}
\usage{
fromCharToFactorNum(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a numeric factor variable
}
\description{
This function takes in a character vector, and returns a numeric factor vector.
}
