% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{getAllProfs}
\alias{getAllProfs}
\title{Create a data frame of all possible profiles}
\usage{
getAllProfs(attributes, idColumnName, dataFrame, replaceNames = TRUE)
}
\arguments{
\item{attributes}{list of attributes}

\item{idColumnName}{a string of the column that identifies respondents}

\item{dataFrame}{the data frame where the attributes came from}

\item{replaceNames}{a boolean to tell the function to replace the colnames}
}
\value{
a dataframe of all possible profiles
}
\description{
This function takes in the attributes of a data frame and returns
a data frame of all of the possible profiles.
}
