% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{getLevels}
\alias{getLevels}
\title{Get a data frame of levels in order as they appear}
\usage{
getLevels(attributes, dataFrame)
}
\arguments{
\item{attributes}{a list of attributes}

\item{dataFrame}{the data frame where the attributes came from}
}
\value{
a data frame of the levels
}
\description{
This function identifies the levels of each attribute provided
in the order they appear in the data frame.
}
