% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{getPredictedPrefs}
\alias{getPredictedPrefs}
\title{Predict the missing values from all possible preferences}
\usage{
getPredictedPrefs(
  attributes,
  allProfs,
  idColumnName,
  answerColumnName,
  dataFrameExp,
  dataframeSurvey
)
}
\arguments{
\item{attributes}{list of attributes}

\item{allProfs}{a dataframe of all possible profiles for each respondent}

\item{idColumnName}{a string of the column that identifies respondents}

\item{answerColumnName}{a string of the column that identifies the survey responses}

\item{dataFrameExp}{a data frame of the experiment data}

\item{dataFrameSurvey}{a data frame of the survey data}
}
\value{
a data frame with the existing responses and predicted responses for each profile from each individual
}
\description{
This creates a random forest model based on the data frames provided, and predicts the responses
for each profile for each respondent. This function has only been tested on the data provided.
}
