% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{getPrefs}
\alias{getPrefs}
\title{Get a data frame of the preferences for each profile}
\usage{
getPrefs(attributes, profs, idColumnName, answerColumnName, dataFrame)
}
\arguments{
\item{attributes}{list of attributes}

\item{profs}{a data frame of the profiles that exist in the data frame provided}

\item{idColumnName}{a string of the column that identifies respondents}

\item{answerColumnName}{a string of the column that identifies the survey responses}

\item{dataFrame}{the data frame where the attributes and profs came from}
}
\value{
a data frame of the preferences for each profile that exists in the data frame
}
\description{
This function creates a data frame with the preferences for each profile.
}
