% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{getProfs}
\alias{getProfs}
\title{Create data frame of unique profiles in a dataset}
\usage{
getProfs(attributes, dataFrame)
}
\arguments{
\item{attributes}{a list of attributes}

\item{dataFrame}{the data frame where the attributes came from}
}
\value{
a data frame of the profiles
}
\description{
This function takes in a list of attributes and a dataframe and returns a data frame of
the existing profiles in the dataset.
}
